fix: recover react query example, contexts
