+++ apps/liff/src/components/EventCard.tsx
