--- apps/web/eslint.config.js
@@ -1,4 +1,0 @@
-import { nextJsConfig } from '@sommhai/eslint-config/next-js';
-
-/** @type {import("eslint").Linter.Config} */
-export default nextJsConfig;
