--- apps/web/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "web",
-  "version": "1.0.0",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "dev": "next dev --port 3000 --turbopack",
-    "build": "next build",
-    "start": "next start",
-    "lint": "eslint \"src/**/*.ts*\"",
-    "check-types": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@sommhai/api-contract": "workspace:*",
-    "@sommhai/eslint-config": "workspace:*",
-    "@sommhai/shared-type": "workspace:*",
-    "@sommhai/tailwind-config": "workspace:*",
-    "@sommhai/typescript-config": "workspace:*",
-    "@sommhai/ui": "workspace:*",
-    "@tanstack/react-query": "^5.66.0",
-    "@ts-rest/core": "^3.51.0",
-    "@ts-rest/react-query": "^3.51.0",
-    "axios": "^1.7.9",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
-    "lucide-react": "^0.475.0",
-    "next": "^15.1.6",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "tailwind-merge": "^3.0.1",
-    "tailwindcss-animate": "^1.0.7"
-  },
-  "devDependencies": {
-    "@next/eslint-plugin-next": "^15.1.6",
-    "@types/node": "^22.12.0",
-    "@types/react": "^19.0.7",
-    "@types/react-dom": "^19.0.3",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^9.19.0",
-    "postcss": "^8.5.1",
-    "tailwindcss": "^3.4.17",
-    "typescript": "5.7.3"
-  }
-}
