--- apps/web/src/app/layout.tsx
@@ -1,20 +1,0 @@
-import './globals.css';
-import '@sommhai/ui/styles.css';
-
-import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
-
-const inter = Inter({ subsets: ['latin'] });
-
-export const metadata: Metadata = {
-  title: 'Create Turborepo',
-  description: 'Generated by create turbo',
-};
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <html lang='en'>
-      <body className={inter.className}>{children}</body>
-    </html>
-  );
-}
