--- apps/web/src/env.ts
@@ -1,1 +1,0 @@
-export const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL || '';
