--- apps/web/tailwind.config.ts
@@ -1,11 +1,0 @@
-// tailwind config is required for editor support
-
-import sharedConfig from '@sommhai/tailwind-config';
-import type { Config } from 'tailwindcss';
-
-const config: Pick<Config, 'content' | 'presets'> = {
-  content: ['./src/**/*.tsx'],
-  presets: [sharedConfig],
-};
-
-export default config;
