--- apps/web/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "@sommhai/typescript-config/nextjs.json",
-  "compilerOptions": {
-    "plugins": [
-      {
-        "name": "next"
-      }
-    ],
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["src/*"],
-      "@ui/*": ["@sommhai/ui/*"]
-    }
-  },
-  "include": ["**/*.ts", "**/*.tsx", "next-env.d.ts", "next.config.ts", ".next/types/**/*.ts"],
-  "exclude": ["node_modules"]
-}
